feat(tmux): add a bit more smooth tmux interactins about creating and killing panes
This change is making it so weather the ctrl key is held down or not for
creating or killing panes/windows should just work, and there is nothing
specific, where e.g. you need to lift ctrl to split the pane, you can
just keep the ctrl pressed while trying to split the pane.